--- common/app/com/commercetools/sunrise/common/controllers/WithOverridablePageContent.java
@@ -1,5 +1,0 @@
-package com.commercetools.sunrise.common.controllers;
-
-public interface WithOverridablePageContent<T> {
-    T createPageContent();
-}
