--- common/app/com/commercetools/sunrise/common/forms/AddressFormBean.java
@@ -1,89 +1,0 @@
-package com.commercetools.sunrise.common.forms;
-
-public class AddressFormBean extends FormBean {
-
-    private String firstName;
-    private String lastName;
-    private String streetName;
-    private String additionalStreetInfo;
-    private String city;
-    private String postalCode;
-    private String region;
-    private String phone;
-    private String email;
-
-    public AddressFormBean() {
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(final String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(final String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getStreetName() {
-        return streetName;
-    }
-
-    public void setStreetName(final String streetName) {
-        this.streetName = streetName;
-    }
-
-    public String getAdditionalStreetInfo() {
-        return additionalStreetInfo;
-    }
-
-    public void setAdditionalStreetInfo(final String additionalStreetInfo) {
-        this.additionalStreetInfo = additionalStreetInfo;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(final String city) {
-        this.city = city;
-    }
-
-    public String getPostalCode() {
-        return postalCode;
-    }
-
-    public void setPostalCode(final String postalCode) {
-        this.postalCode = postalCode;
-    }
-
-    public String getRegion() {
-        return region;
-    }
-
-    public void setRegion(final String region) {
-        this.region = region;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(final String phone) {
-        this.phone = phone;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(final String email) {
-        this.email = email;
-    }
-}
