+++ shopping-cart/app/com/commercetools/sunrise/shoppingcart/checkout/address/CheckoutAddressFormSettings.java
@@ -1,8 +1,8 @@
 package com.commercetools.sunrise.shoppingcart.checkout.address;
 
-import com.commercetools.sunrise.common.models.CountryFormFieldBean;
+import com.commercetools.sunrise.common.forms.CountryFormFieldBean;
 import com.commercetools.sunrise.common.models.ModelBean;
-import com.commercetools.sunrise.common.models.TitleFormFieldBean;
+import com.commercetools.sunrise.common.forms.TitleFormFieldBean;
 
 public class CheckoutAddressFormSettings extends ModelBean {
